test: 테스트 코드 리팩토링
* test: 테스트코드 리팩토링
* docs: 구현한 기능
